Refactor ObjectDataSet
